% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_route_unique.R
\name{add_route_unique}
\alias{add_route_unique}
\title{Add a column to a BBS table identifying unique routes survey-wide}
\usage{
add_route_unique(df, col_name = "route_unique")
}
\arguments{
\item{df}{A \code{data.frame} created by a \code{bbs_*} function that
includes columns \code{country_num}, \code{state_num}, and \code{route}.}

\item{col_name}{Name for the new column. Defaults to "route_unique".}
}
\value{
\code{df} with an additional column that identifies unique routes
survey-wide. The additional column is called \code{name} and is of class
\code{integer}.
}
\description{
The \code{route} column in a variety of BBS tables only identifies unique
routes within states. This function creates route identifiers that are unique
survey-wide by pasting together columns \code{country_num}, \code{state_num},
and \code{route}.
}
\examples{
\dontrun{

# load the metadata table 'weather'
weather <- bbs_meta_weather(bbs_dir = ".")

# add a column identifying unique routes
weather <- add_route_unique(weather)
}

}
\references{
Pardieck, K.L., D.J. Ziolkowski Jr., M. Lutmerding and M.-A.R.
  Hudson. 2018. North American Breeding Bird Survey Dataset 1966-2017,
  version 2017.0. U.S. Geological Survey, Patuxent Wildlife Research Center.
  \url{https://doi.org/10.5066/F76972V8}
}
\author{
Bob O'Hara

Patrick Barks <patrick.barks@gmail.com>
}
