% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_build_50.R
\name{bbs_build_50}
\alias{bbs_build_50}
\title{Assemble North American Breeding Bird Survey 50-stop count data}
\usage{
bbs_build_50(bbs_dir)
}
\arguments{
\item{bbs_dir}{Local directory from which to load BBS data.

  Must be an existing directory with the same structure as the USGS BBS
  directory (ftp://ftpext.usgs.gov/pub/er/md/laurel/BBS/DataFiles/).}
}
\value{
A \code{data.frame} with the following columns:
  \item{route_data_id}{integer code for unique combinations of country_num,
  state, route, rpid, and year}
  \item{country_num}{integer code for country: 124 (Canada), 484 (Mexico), or
  840 (United States)}
  \item{state_num}{integer code for state/province/territory (see BBS file
  \emph{RegionCodes.txt})}
  \item{route}{integer code for survey route (unique within states)}
  \item{rpid}{integer Run Protocol Identification ID (see BBS file
  \emph{RunProtocolID.txt})}
  \item{year}{integer survey year}
  \item{aou}{integer species code from American Ornithological Union}
  \item{stop_1 ... stop_50}{total number of individuals of the species
  recorded on the given stop}
}
\description{
Assemble North American Breeding Bird Survey (BBS) 50-stop count data
}
\examples{

}
\references{
Pardieck, K.L., D.J. Ziolkowski Jr., M. Lutmerding and M.-A.R.
  Hudson. 2018. North American Breeding Bird Survey Dataset 1966-2017,
  version 2017.0. U.S. Geological Survey, Patuxent Wildlife Research Center.
  \url{https://doi.org/10.5066/F76972V8}
}
\author{
Bob O'Hara

Patrick Barks <patrick.barks@gmail.com>
}
