% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_files_10.R
\name{bbs_files_10}
\alias{bbs_files_10}
\title{Get the names of state-specific zip files containing 10-stop count data from
README.txt}
\usage{
bbs_files_10(bbs_dir = NULL)
}
\arguments{
\item{bbs_dir}{Directory from which to get data. Defaults to the USGS FTP
directory for the most recent BBS release. May alternatively be a path to a
local directory, or ftp address for an older BBS release.}
}
\value{
\code{data.frame} with the following columns:
  \item{state_name}{name of state/province/territory}
  \item{ten_stop_file}{name of zip file with 10-stop survey data}
}
\description{
Get the names of state-specific zip files containing 10-stop count data from
the BBS file \emph{README.txt}.
}
\details{
Note that not all regions have a zip file. As of 2017 there is no
  data for Mexico, Puerto Rico, or Washington D.C.
}
\examples{

}
\references{
Pardieck, K.L., D.J. Ziolkowski Jr., M. Lutmerding and M.-A.R.
  Hudson. 2018. North American Breeding Bird Survey Dataset 1966-2017,
  version 2017.0. U.S. Geological Survey, Patuxent Wildlife Research Center.
  \url{https://doi.org/10.5066/F76972V8}
}
\author{
Bob O'Hara

Patrick Barks <patrick.barks@gmail.com>
}
