% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_meta_regions.R
\name{bbs_meta_regions}
\alias{bbs_meta_regions}
\title{Read BBS meta-data}
\usage{
bbs_meta_regions(bbs_dir = NULL, zip_files = FALSE)
}
\arguments{
\item{bbs_dir}{Directory from which to get data. Defaults to the USGS FTP
directory for the most recent BBS release. May alternatively be a path to a
local directory, or ftp address for an older BBS release.}

\item{zip_files}{Logical: should the names of the zip files for the 10- and
50-stop data be added? Defaults to FALSE.}
}
\value{
\code{data.frame} with the following columns:
  \item{country_num}{Integer code for country}
  \item{country_name}{Name of country}
  \item{state_num}{Integer code for state/province/territory}
  \item{state_name}{Name of state/province/territory}
  \item{ten_stop_file}{Name of zip file with 10-stop survey data}
}
\description{
Read in list of states/provinces/territories, and also names of zip files
where the 10 stop data is kept.
}
\details{
This is meta-data collated from the full database. Note that not all
  regions have a zip file: in particular, there is no Mexican data, or data
  from Puerto Rico or Washington D.C.
}
\examples{
regions <- bbs_meta_regions()

}
\references{
Sauer, J. R., J. E. Hines, J. E. Fallon, K. L. Pardieck, D. J.
  Ziolkowski, Jr., and W. A. Link. 2014. The North American Breeding Bird
  Survey, Results and Analysis 1966 - 2012. Version 02.19.2014 USGS Patuxent
  Wildlife Research Center, Laurel, MD
}
\author{
Bob O'Hara

Patrick Barks <patrick.barks@gmail.com>
}
