% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_meta_strata.R
\name{bbs_meta_strata}
\alias{bbs_meta_strata}
\title{Get geographic strata data for North American Breeding Bird Survey}
\usage{
bbs_meta_strata(bbs_dir = NULL)
}
\arguments{
\item{bbs_dir}{Directory from which to get data. Defaults to the USGS FTP
directory for the most recent BBS release. May alternatively be a path to a
local directory, or ftp address for an older BBS release.}
}
\value{
\code{data.frame} with the following columns:
  \item{stratum}{Integer code for stratum}
  \item{stratum_name}{Stratum name in English}
  \item{stratum_name_french}{Stratum name in French}
  \item{stratum_name_spanish}{Stratum name in Spanish}
}
\description{
Get names and ID numbers of geographic strata in the North American Breeding
Bird Survey (BBS). This data comes from the BBS file 'BBSStrata.txt'.
}
\examples{
strata <- bbs_meta_strata()

}
\references{
Sauer, J. R., J. E. Hines, J. E. Fallon, K. L. Pardieck, D. J.
  Ziolkowski, Jr., and W. A. Link. 2014. The North American Breeding Bird
  Survey, Results and Analysis 1966 - 2012. Version 02.19.2014 USGS Patuxent
  Wildlife Research Center, Laurel, MD
}
\author{
Bob O'Hara

Patrick Barks <patrick.barks@gmail.com>
}
