% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_unzip.R
\name{csv_unzip}
\alias{csv_unzip}
\title{Extract a csv file from a zipped archive of the same name}
\usage{
csv_unzip(zip_path)
}
\arguments{
\item{zip_path}{Path to zipped archive}
}
\value{
A \code{data.frame} returned by \code{read.csv}
}
\description{
Extract a csv file from a zipped archive of the same name (except for the
file extension).
}
\examples{

}
\references{
Sauer, J. R., J. E. Hines, J. E. Fallon, K. L. Pardieck, D. J.
  Ziolkowski, Jr., and W. A. Link. 2014. The North American Breeding Bird
  Survey, Results and Analysis 1966 - 2012. Version 02.19.2014 USGS Patuxent
  Wildlife Research Center, Laurel, MD
}
\author{
Bob O'Hara

Patrick Barks <patrick.barks@gmail.com>
}
