% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRegions.R
\name{GetRegions}
\alias{GetRegions}
\title{Read BBS meta-data}
\usage{
GetRegions(bbs_dir = NULL, ZipFiles = FALSE)
}
\arguments{
\item{bbs_dir}{Directory from which to get data. Defaults to the USGS FTP
directory for the most recent BBS release. May alternatively be a path to a
local directory, or ftp address for an older BBS release.}

\item{ZipFiles}{Logical: should the names of the zip files for the 10- and
50-stop data be added? Defaults to FALSE.}
}
\value{
\code{data.frame} with the following columns:
  \item{countrynum}{Country code}
  \item{RegionCode}{Region code}
  \item{State/Prov/TerrName}{Name of (US) state, (Canadian) province or
  (Mexican) territory}
  \item{CountryName}{Name of country}
  \item{FileName10stop}{Name of zip file with 10 stop data}
  \item{FileName50stop}{Name of zip file with 50 stop data}
}
\description{
Read in list of states/provinces/territories, and also names of zip files
where the 10 stop data is kept.
}
\details{
This is meta-data collated from the full database. Nore that not all
  regions have a zip file: in particular, there is no Mexican data, or data
  from Puerto Rico or Washington D.C.
}
\examples{
Regions <- GetRegions()

}
\references{
Sauer, J. R., J. E. Hines, J. E. Fallon, K. L. Pardieck, D. J.
  Ziolkowski, Jr., and W. A. Link. 2014. The North American Breeding Bird
  Survey, Results and Analysis 1966 - 2012. Version 02.19.2014 USGS Patuxent
  Wildlife Research Center, Laurel, MD
}
\author{
Bob O'Hara
}
