% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSpNames.R
\name{GetSpNames}
\alias{GetSpNames}
\title{Read BBS meta-data}
\usage{

  GetSpNames(Dir = "ftp://ftpext.usgs.gov/pub/er/md/laurel/BBS/DataFiles/")
}
\arguments{
\item{Dir}{Directory to get data. Defaults to
ftp://ftpext.usgs.gov/pub/er/md/laurel/BBS/DataFiles/, the USGS FTP server}
}
\value{
Data frame with these columns:
  \item{Seq}{Phylogenetic sequence number}
  \item{AOU}{American Ornithological Union code number}
  \item{English_Common_Name}{English Common Name}
  \item{French_Common_Name}{French Common Name}
  \item{Spanish_Common_Name}{Spanish Common Name}
  \item{ORDER}{Taxonomic order}
  \item{Family}{Taxonomic family}
  \item{Genus}{Taxonomic genus}
  \item{Species}{Taxonomic species name}
}
\description{
Read in list of species names, from SpeciesList.txt, and then extract list of
where the data is kept
}
\examples{
Species <- GetSpNames()

}
\references{
Sauer, J. R., J. E. Hines, J. E. Fallon, K. L. Pardieck, D. J.
  Ziolkowski, Jr., and W. A. Link. 2014. The North American Breeding Bird
  Survey, Results and Analysis 1966 - 2012. Version 02.19.2014 USGS Patuxent
  Wildlife Research Center, Laurel, MD
}
\author{
Bob O'Hara
}
