% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSppCounts.R
\name{GetSppCounts}
\alias{GetSppCounts}
\title{GetSppCounts}
\usage{
GetSppCounts(count = bbs, AOU, run_atrb = weather, route_atrb = routes,
  years = seq(from = 1997, to = 2014), statenum = NULL, countrynum = NULL)
}
\arguments{
\item{count}{Data frame containing the raw BBS counts obtained from the function `GetRouteData()`}

\item{AOU}{Numeric AOU code for the species of interest}

\item{years}{Optional vector containing the years of interest (default in full study period)}

\item{statenum}{Optional numeric vector containing the states of interest (see BBS website for code values)}

\item{countrynum}{Optional numeric vector containing the countries of interest (840 = US, 124 = Canada, 484 = Mexico)}

\item{weather}{Data frame containing the weather data obtained from the function `GetWeather()`}

\item{routes}{Data frame containing the route data obtained from the function `GetRoutes()`}
}
\value{
A .csv file containing the following fields:

routeID The unique 8 digit route ID for each route

Year The year that the count was conducted

aou The numeric code for the focal species

countN Total individuals of the focal species recorded at stop N (50-stop data) or N-9:N (10-stop data)

Latitude The latitude for the route

Longitude The longitude for the route

Stratum The BBS stratum for the route

BCR The Bird Conservation Region for the route
}
\description{
Filter raw BBS counts by species, add 0 counts, and latitude/longitude
}

