% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetUnzip.R
\name{GetUnzip}
\alias{GetUnzip}
\title{Function to extract csv data from a zipped archive}
\usage{
GetUnzip(ZipName, FileName)
}
\arguments{
\item{ZipName}{name of zipped archive}

\item{FileName}{name of file in zipped archive}
}
\value{
A \code{data.frame} returned by \code{read.csv}
}
\description{
Code adapted from
http://stackoverflow.com/questions/3053833/using-r-to-download-zipped-data-file-extract-and-import-data
}
\examples{
## Download woodpecker data from BBS
Rep <- "ftp://ftpext.usgs.gov/pub/er/md/laurel/BBS/DataFiles/"
dat <- GetUnzip(ZipName = paste0(Rep, "States/Florida.zip"),
                FileName="Florida.csv")

\dontrun{
  ## Download from a local repository
  ##   Use your own LocalRepos
  LocalRepos <- "~/Dropbox/"
  dat=GetUnzip(ZipName=paste0(LocalRepos, 
  "BBS/DataFiles/Species/TrogWoodp.zip"), 
  FileName="TrogWoodp.csv")
}

}
\references{
Sauer, J. R., J. E. Hines, J. E. Fallon, K. L. Pardieck, D. J.
  Ziolkowski, Jr., and W. A. Link. 2014. The North American Breeding Bird
  Survey, Results and Analysis 1966 - 2012. Version 02.19.2014 USGS Patuxent
  Wildlife Research Center, Laurel, MD
}
\author{
Bob O'Hara
}
