\name{rBBS-package}
\alias{rBBS-package}
\alias{rBBS}
\docType{package}
\title{
Import data from the USGS North American Breeding Bird Survey data base
}
\description{
A package to import the USGS's BBS data from their ftp server. 
}
\details{
\tabular{ll}{
Package: \tab rBBS\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-01-09\cr
License: \tab GPL-3.0\cr
}
}
\author{
Bob O'Hara

Maintainer: Bob O'Hara <bohara@senckenberg.de>
}
\references{
Sauer, J. R., J. E. Hines, J. E. Fallon, K. L. Pardieck, D. J. Ziolkowski, Jr., and W. A. Link. 2014. The North American Breeding Bird Survey, Results and Analysis 1966 - 2012. Version 02.19.2014 USGS Patuxent Wildlife Research Center, Laurel, MD

http://www.mbr-pwrc.usgs.gov/bbs/bbs.html
}
\keyword{ package }
\examples{
## Would love to run this but function disappears
## CollaredDove <- Get10RouteData(ColDoveAOU, weather = NULL, routes = NULL, 
##                      year=c(1995,2000,2005, 2010))
}
