% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rBBS_package.R
\docType{package}
\name{rBBS}
\alias{rBBS}
\alias{rBBS-package}
\title{rBBS: Access the North American Breeding Bird Survey (BBS) in R}
\description{
An R package to work with data from the North American Breeding Bird Survey
(BBS).
}
\details{
The BBS is a cooperative effort between the United States Geological Survey
(USGS) and Environment and Climate Change Canada's Canadian Wildlife Service
(CWS). See more information at \url{https://www.pwrc.usgs.gov/bbs/}, and see
also the terms of use for BBS data at
\url{https://www.pwrc.usgs.gov/BBS/RawData/}.
}
\references{
Pardieck, K.L., D.J. Ziolkowski Jr., M. Lutmerding and M.-A.R.
  Hudson. 2018. North American Breeding Bird Survey Dataset 1966-2017,
  version 2017.0. U.S. Geological Survey, Patuxent Wildlife Research Center.
  \url{https://doi.org/10.5066/F76972V8}
}
\author{
Bob O'Hara

Patrick Barks <patrick.barks@gmail.com>
}
