% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_build_50.R
\name{bbs_build_50}
\alias{bbs_build_50}
\title{Build data frame with North American Breeding Bird Survey 10-stop data}
\usage{
bbs_build_50(bbs_dir)
}
\arguments{
\item{bbs_dir}{Directory from which to load data. Must be a path to an
existing local directory.}
}
\description{
Build data frame with North American Breeding Bird Survey 10-stop data
}
\examples{

}
\references{
Sauer, J. R., J. E. Hines, J. E. Fallon, K. L. Pardieck, D. J.
  Ziolkowski, Jr., and W. A. Link. 2014. The North American Breeding Bird
  Survey, Results and Analysis 1966 - 2012. Version 02.19.2014 USGS Patuxent
  Wildlife Research Center, Laurel, MD
}
\author{
Patrick Barks <patrick.barks@gmail.com>
}
