% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_download.R
\name{bbs_download}
\alias{bbs_download}
\title{Download North American Breeding Bird Survey data via FTP}
\usage{
bbs_download(dest, bbs_dir = NULL, stateprovs = NULL, meta = TRUE,
  ten_stop = TRUE, fifty_stop = FALSE, migrant = FALSE)
}
\arguments{
\item{dest}{Path to an existing local directory to download files to.}

\item{bbs_dir}{Directory from which to get data. Defaults to the USGS FTP
directory for the most recent BBS release. May alternatively be an FTP
address for an older BBS release.}

\item{stateprovs}{Vector of countries/states/provinces to download data for.
Used only if \code{ten_stop == TRUE}. Defaults to all available regions.
Case-insensitive. See Details.}

\item{meta}{Download top-level metadata files (T/F)? Default is \code{TRUE}.}

\item{ten_stop}{Download 10-stop data (T/F)? Default is \code{TRUE}.}

\item{fifty_stop}{Download 50-stop data (T/F)? Default is \code{FALSE}.}

\item{migrant}{Download data for migrant non-breeders (T/F)? Default is
\code{FALSE}.}
}
\description{
Download North American Breeding Bird Survey (BBS) data via FTP.
}
\details{
Argument \code{stateprov} may include names of countries ('Canada'
  or 'United States') and/or states/provinces/territories. Different
  geographic levels may be combined; e.g. \code{stateprov = c('Canada',
  'Montana', 'North Dakota')} will download data for all Canadian
  provinces/territories plus Montana and North Dakota.
}
\examples{
\dontrun{
# download metadata and 10-stop data for Alaska and all of Canada
bbs_download(dest = '~/bbs/', stateprovs = c('Alaska', 'Canada'))

# download metadata and 10-stop data for Gulf Coast states
gulf_states <- c('Alabama', 'Florida', 'Louisiana', 'Mississippi', 'Texas')
bbs_download(dest = '~/bbs/', stateprovs = gulf_states)

# download metadata and 50-stop data for all regions
bbs_download(dest = '~/bbs/', ten_stop = FALSE, fifty_stop = TRUE)
}

}
\references{
Pardieck, K.L., D.J. Ziolkowski Jr., M. Lutmerding and M.-A.R.
  Hudson. 2018. North American Breeding Bird Survey Dataset 1966-2017,
  version 2017.0. U.S. Geological Survey, Patuxent Wildlife Research Center.
  \url{https://doi.org/10.5066/F76972V8}
}
\author{
Bob O'Hara

Patrick Barks <patrick.barks@gmail.com>
}
