% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_download.R
\name{bbs_download}
\alias{bbs_download}
\title{Download North American Breeding Bird Survey data via FTP}
\usage{
bbs_download(dest, bbs_dir = bbs_ftp(), countries = NULL,
  states = NULL, meta = TRUE, ten_stop = TRUE, fifty_stop = FALSE,
  overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{dest}{Path to an existing local directory to download files to.}

\item{bbs_dir}{Directory from which to get data. Defaults to the USGS FTP
directory for the most recent BBS release. May alternatively be an FTP
address for an older BBS release.}

\item{countries}{Vector of countries to download 10- and/or 50-stop data for
(e.g. \code{c("Canada", "United States")}). Defaults to all available.
Case-insensitive. See Details.}

\item{states}{Vector of states/provinces/territories to download 10- and/or
50-stop data for. Defaults to all available. Case-insensitive. See Details.}

\item{meta}{Download top-level metadata files? Default is \code{TRUE}.}

\item{ten_stop}{Download 10-stop data? Default is \code{TRUE}.}

\item{fifty_stop}{Download 50-stop data? Default is \code{FALSE}.}

\item{overwrite}{Overwrite files that already exist? Default is \code{FALSE}.}

\item{verbose}{List files that were not re-downloaded because they already
exist? Default is \code{TRUE}. Only applies if \code{overwrite = FALSE}.}
}
\description{
Download North American Breeding Bird Survey (BBS) data via FTP.
}
\details{
For names to use in arguments \code{countries} and \code{states} see
\code{\link{bbs_meta_regions}} or BBS file \emph{RegionCodes.txt}.

Note that country/state subsets are additive, so specifying

\code{
bbs_download(".", countries = "Canada", states = "Montana")
}

will download data for all Canadian provinces/territories plus the state of
Montana. The following lines will both download data for all American states:

\code{bbs_download(".", countries = "United States", states = "Florida")} \cr
\code{bbs_download(".", countries = "United States")}
}
\examples{
\dontrun{

# download metadata and 10-stop data for Alaska and all of Canada
bbs_download(dest = ".", countries = "Canada", states = "Alaska")

# download metadata and 10-stop data for Gulf Coast states
gulf_states <- c("Alabama", "Florida", "Louisiana", "Mississippi", "Texas")
bbs_download(dest = ".", states = gulf_states)

# download metadata and 50-stop data for all regions
bbs_download(dest = ".", ten_stop = FALSE, fifty_stop = TRUE)
}

}
\references{
Pardieck, K.L., D.J. Ziolkowski Jr., M. Lutmerding and M.-A.R.
  Hudson. 2018. North American Breeding Bird Survey Dataset 1966-2017,
  version 2017.0. U.S. Geological Survey, Patuxent Wildlife Research Center.
  \url{https://doi.org/10.5066/F76972V8}
}
\author{
Bob O'Hara

Patrick Barks <patrick.barks@gmail.com>
}
